refactor: move files for tests
