+++ src/routes/error-page/Error.test.tsx
@@ -1,6 +1,6 @@
 import { test } from 'vitest';
 import ErrorPage from './Error';
-import renderWithRouterAndRedux from '../../__tests__/render-with-router';
+import renderWithRouterAndRedux from '../../../tests/render-with-router-redux';
 
 test('About screen should be rendered', () => {
   const { getByText } = renderWithRouterAndRedux(<ErrorPage />);
