+++ tsconfig.json
@@ -18,6 +18,6 @@
     "noEmit": true,
     "jsx": "react-jsx"
   },
-  "include": ["src"],
+  "include": ["src", "tests/render-with-router-redux.tsx"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
